% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_vblock}
\alias{get_vblock}

\title{
  Parse BGEN variant block
}
\description{
  Retrieves all variant information from the BGEN variant block. 
}

\arguments{
\item{none}{}
}

\details{
The BGEN file must be opened first using \code{open_bgen} to use this function. The pointer will reset back to the beginning of the very first variant block so that other functions may be used after. If no other functions will be used after, close the BGEN file with \code{close_bgen()}.
}
\value{
  A data.frame containing information from the BGEN file. The column names are described below. Columns \code{A1} and \code{A2} will be NA if \code{ALLELES} is not 2.
  \item{SNPID}{The variant identifier.}
  \item{RSID}{The reference snp identifier.}
  \item{CHR}{The chromosome of the variant.}
  \item{POS}{The position of the variant.}
  \item{ALLELES}{The number of alleles.}
  \item{A1}{The first allele in the BGEN file.}
  \item{A2}{The second allele in the BGEN file.}
}
\examples{
  bgenInfo <- open_bgen("my_bgenFile.bgen")
  variantInfo <- get_vblock()
  close_bgen()
}
