% Generated by roxygen2 (4.0.0): do not edit by hand
\name{open_bgen}
\alias{open_bgen}

\title{
  Open BGEN file
}
\description{
  Creates a global C stream to the input BGEN file (see Details) and returns values from the header block and sample identifier block of the BGEN file.
}

\arguments{
\item{bgenFile}{Path to the BGEN file.}

\item{asSampleFile}{Logical value indicating whether the sample identifiers should be returned as a dataframe in .sample file format. If FALSE, a vector of the sample identifiers is returned.}
}
\details{
The C stream is a global object that should be remained open to use other functions in this package. Use \code{close_bgen()} to close the BGEN file safely.
}
\value{
A list is returned containing information from the header block and sample identifier block of the BGEN file.
}
\examples{
bgenInfo <- open_bgen("my_bgenFile.bgen")
}
